Move views into views package
