+++ src/main/java/com/minecolonies/blockout/controls/TextField.java
@@ -2,7 +2,7 @@ package com.minecolonies.blockout.controls;
 
 import com.minecolonies.blockout.Pane;
 import com.minecolonies.blockout.PaneParams;
-import com.minecolonies.blockout.View;
+import com.minecolonies.blockout.views.View;
 import net.minecraft.client.gui.GuiScreen;
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.client.renderer.Tessellator;
