+++ src/main/java/com/minecolonies/blockout/views/ScrollingContainer.java
@@ -2,7 +2,6 @@ package com.minecolonies.blockout.views;
 
 import com.minecolonies.blockout.Pane;
 import com.minecolonies.blockout.PaneParams;
-import com.minecolonies.blockout.View;
 import net.minecraft.client.renderer.GlStateManager;
 import org.jetbrains.annotations.NotNull;
