+++ src/main/java/com/minecolonies/coremod/client/gui/DialogDoneCancel.java
@@ -1,7 +1,7 @@
 package com.minecolonies.coremod.client.gui;
 
 import com.minecolonies.blockout.Loader;
-import com.minecolonies.blockout.OverlayView;
+import com.minecolonies.blockout.views.OverlayView;
 import com.minecolonies.blockout.controls.Button;
 import com.minecolonies.blockout.controls.Label;
 import com.minecolonies.blockout.controls.Text;
